Added support kernel_recovery option
This flashes u-boot, kernel and initramfs and copies
the kernel, modules and firmware over to rootfs. Note
that it will possibly overwrite things installed into
the rootfs, so those need to be reinstalled with fixed
versions.

Note we may want to extract the kernel modules from
the initramfs and install them and the kernel from the
installer script instead.