Update to u-boot v2014.10
This gets us a bit closer to being able to send patches
upstream at some point and we can start using the armv7
and omap3 generic headers. And we can start using the
faster booting straight from SPL.

The biggest change is to start using the armv7 default
load address 0x80800000 instead of 0x80008000 that causes
changes all over the place for addressses. Presumably
that's not going to change any more in u-boot.